Move Fullscreen enum into platforms
Needed if we intend to split these into separate crates.
